requirements_dev.txt
# Runtime dependencies
Django>=3.0
django-mptt>=0.11.0

# Testing
coverage
flake8
tox

# Documentation
Sphinx
sphinx-rtd-theme

# Other tools
bumpversion==0.5.3
wheel==0.29.0
